refactor(order-service): relocate redis request limiter
