+++ apps/order-service/src/app/order-workflow/infra/auth/request-cooldown/request-cooldown-config.token.ts
@@ -1,4 +1,4 @@
-import type { RequestCooldownConfig } from '../config/request-cooldown.config';
+import type { RequestCooldownConfig } from '../../config/request-cooldown.config';
 
 export const REQUEST_COOLDOWN_CONFIG = 'REQUEST_COOLDOWN_CONFIG';
